Split off CPUID functionality
