+++ cpuid.h
@@ -1,0 +1,4 @@
+#include <stdbool.h>
+#include <stdint.h>
+
+bool has_rdrand(void);
